fix: Changed to low precission for uniform in grid3D layer. (#2146)

+++ typescript/packages/subsurface-viewer/src/layers/grid3d/vertex.glsl.ts
@@ -7,7 +7,7 @@ attribute vec3 positions;
 attribute float properties;
 attribute vec3 normals;
 
-uniform int coloringMode;
+uniform lowp int coloringMode;
 
 // Outputs to fragment shader
 out vec3 cameraPosition;
